+++ NEWS.md
@@ -1,7 +1,8 @@
 # pkgdiff 1.0
 
-* Added logo
-* Added `pkg_report()` function for interactive use.
+* Added logo.
+* Added `pkg_summary()` function for interactive use.
+* Fixed bugs on `pkg_info()` when requesting info on a Base R package. 
 * Upgraded life cycle to "stable".
 
 # pkgdiff 0.2.1